+++ src/asp/Core/PointUtils.h
@@ -424,20 +424,6 @@ vw::ImageViewRef<PixelT> form_point_cloud_composite(std::vector<std::string> con
 // Find the average longitude for a given point image with lon, lat, height values
 double find_avg_lon(vw::ImageViewRef<vw::Vector3> const& point_image);
 
-
-// Get a generous estimate of the bounding box of the current set
-// while excluding outliers
-void estimate_points_bdbox(vw::ImageViewRef<vw::Vector3> const& proj_points,
-                           vw::ImageViewRef<double> const& error_image,
-                           vw::Vector2 const& remove_outliers_params,
-                           double estim_max_error,
-                           vw::BBox3 & estim_bdbox);
-  
-double estim_max_tri_error_and_proj_box(vw::ImageViewRef<vw::Vector3> const& proj_points,
-                                        vw::ImageViewRef<double> const& error_image,
-                                        vw::Vector2 const& remove_outliers_params,
-                                        vw::BBox3 & estim_proj_box);
-  
 std::string get_cloud_type(std::string const& file_name);
   
 // Find the number of channels in the point clouds.